image_process/requirements.txt
labelImg
pillow
imgaug
tqdm
xml